BUG: Fix unnecessary warning of missing subject hierarchy item
There are many reasons why a subject hierarchy item is not created for a node so there is not a reason to throw a critical warning regarding the situation.

Nodes defined to be hidden from editors were being unnecessarily warned by the "Unable to find subject hierarchy item by data node" log message because it is expected for them to not be visible and available in the subject hierarchy as they are hidden.

Code that previously produced the unnecessary warning:
```python
slicer.util.selectModule("Markups")
line_node = slicer.vtkMRMLMarkupsLineNode()
line_node.SetHideFromEditors(True)
slicer.mrmlScene.AddNode(line_node)
```

Co-authored-by: Andras Lasso <lasso@queensu.ca>