BUG: Fix loading of multi-file volumes in SampleData module
When loading sample datasets consisting of multiple files, such as
`DTIVolume.nhdr` and `DTIVolume.raw.gz`, the dataset failed to load.
This was caused by automatically generating a node name for the raw
sidecar file, which was not meant to be loaded into the scene.

Fixed by only generating default node names if none are specified for
any of the files in the sample.

Fixes regression introduced in 89104423af ("BUG: Ensure MRB loading
correctly reports loaded nodes", 2025-01-10).