STYLE: Update python scripts to fix RUF059 unpacked variable never used
See https://docs.astral.sh/ruff/rules/unused-unpacked-variable/

Adding the prefix underscore allows the variables to still be defined for
easy use in the future to be able to use the variable and understand what
is unpacked at that position.

Changes applied by temporarily updating the `ruff` version to `v0.13.0` and
specifying `--unsafe-fixes` in the `.pre-commit-config.yaml`. Remaining errors
were manually reviewed and fixed

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>