BUG: Fix MIP and MinIP VolumeRendering with SSAO render pass
SSAO pass is not compatible with MIP and MinIP Volume rendering blend
modes.

When the SSAO pass is active and the volume has MIP or MinIP blend
modes, the SSAO pass is removed from the actor's
vtkOpenGLRenderPass::RenderPasses() information. key.

The vtkOpenGLRenderPass::RenderPasses() key is entirely removed if after
SSAO pass removal the key is empty to avoid crashes in the GPU mapper.

Following this FIX, some visual artifacts remain when displaying
geometry inside the volume in MIP / MinIP due to improper SSAO texture
cleanup. These artifacts can only be addressed in the VTK shader code.

Closes #7560