BUG: Fix Undo/Redo in the segment editor logic
Fix Undo/Redo in the segment editor not activated when loading a saved
scene.

Problem was due to not observing the segmentation node modified event
and not synchronizing the segmentation history to the segmentation the
first time it was set.

This fix properly calls the observer reconnect and segmentation history
synchronization methods on set.

Closes #8721