ENH: Add volume reorientation to Crop Volume module
Orientation of the input volume can now be easily fixed in Crop Volume module using the new "Reorient volume" section.

Options are added to tune "Fit to volume" behavior. It is now possible to align the cropped volume's ROI to the world coordinate system (or keep the current ROI orientation).
This is useful if the user wants to align the cropped volume axes to the world coordinate system, for example to make segmentation in anatomical planes easier.

fixes #8415