ENH: Make window/level adjustment available in 3D view for volume rendering
This enhancement allows users to adjust window/level settings directly in the
3D view for volume-rendered images.

In "Adjust window/level" mouse mode, click-and-drag on a volume-rendered image
scales and shifts the scalar opacity and color transfer functions.
Ctrl-click-and-drag adjust the opacity of the transfer function.

If multiple volumes are displayed using volume rendering then the one that the
user clicked on will be adjusted.