ENH: Add support for storing modeling displacement field transforms
Displacement field transform can be defined using a vector field that transforms from the target coordinate system to the source object (a.k.a. resampling transform, because it can be used to transform images by resampling; "from parent" in Slicer naming convention), or that transforms form the source object coordinate system to the target (a.k.a. modeling transform, because it can be used to transform geometric objects by moving their points; "to parent" in Slicer naming convention).

We use ITK readers/writers for storing of transforms in files, and since ITK is an image-oriented toolkit, it always stores resampling transforms in files. Slicer worked around this limitation by introducing special "inverse" transform classes - which specify that those transforms need to be inverted before using them as resampling transform.

However, when displacement field transforms were stored as images. There was no way to specify in the image files that the displacement field should be used for resampling or modeling.

This commit adds the possibility to store displacement fields in NRRD files as either resampling and modeling transform, by introducing a new optional custom field: "displacement field type". This field can be set to "resampling" (default) or "modeling". For other file formats (NIFTI, MetaImage, ...) it is still assumed that the displacement field should be used for resampling.