ENH: Save node attributes in volume and transform sequence files
When a volume sequence is saved in MRML sequence bundle (.seq.mrb) format then all the node attributes of the data nodes in the sequence are preserved.

However, the volume and transform sequence readers, which saved data more efficiently in .seq.nrrd format could not store these attributes.
This commits implements this missing feature.

fixes #7218