ENH: Add nvidia warp GPU grow cut
This adds a Warp GPU implementation of the underlyging GrowCut
algorithm as an option if support is available (Warp with CUDA).
If also offers a CPU version based on Warp, but this may be slower
than the vtkITK version.

On informaal testing the Warp/CUDA version appears to be much faster than
the vtkITK version.  Results appear to be identical.  Performance
will vary based on GPU used.

This code was ported from an earlier implementation based on OpenCL:

https://github.com/pieper/SlicerCL/blob/3d04661ef7f225edf00292a92088db902787a016/GrowCutCL/GrowCutCL.cl.in

The Warp version and changes to the GrowFromSeeds effect were implemented
with the help of Google Gemini.