ENH: Show launcher splashscreen until Slicer is started
Previously, the splashscreen of the launcher was only displayed for 3 seconds, but often the application's splashscreen did not show up for 10-20 seconds.
It was confusing for users to see two splashscreens appearing during Slicer startup, with some time between them when no splashscreen was displayed
(and so users did not know if the application startup was in progress).

This commit makes the application launcher splashscreen displayed until the application can display its own splashscreen.
This is implemented by leveraging the CTK application launcher's new feature, which hides its splashscreen when the launched application writes something on its output.

see #8638