BUG: Fix loading of orientation marker and fonts
Recent changes in application initialization resulted in incorrectly computed paths for orientation marker and font files.

The root cause of the issue was that vtkMRMLApplicationLogic::GetShareDirectory() returns a relative path.

Fixed the issue by properly constructing full path from the relative path. Added a convenience function that makes it easies and less error-prone to get the full path of files and folders in the application share folder.

Co-authored-by: Andras Lasso <lasso@queensu.ca>