BUG: Ensure launcher splash-screen gets hidden on app restart
When the application was restarted, the launcher splash-screen did not get hidden.
This is due to writing to process output is disabled when the application is restarted
and the launcher splash-screen hiding mechanism relies on writing to the process output.

Resolved the issue by not showing the launcher splash-screen when restarting the application.

see #8638