BUG: Fix test assuming extension manager enabled
The test assumed that Slicer_EXTENSIONS_DIRBASENAME is defined,
but it wouldn't be defined if extension manager support were disabled.

This caused problems when trying to build a custom application with

Slicer_BUILD_EXTENSIONMANAGER_SUPPORT OFF
BUILD_TESTING ON

The issue was introduced in 2803638e40fdf3fe7b478277247d097ce264d4e5
which did guard with `#ifdef Slicer_BUILD_EXTENSIONMANAGER_SUPPORT`
in core code, but probably just forgot to do it in this test.