BUG: Fix Qt6 incompatible APIs in tests
QTime does not have start() method anymore in Qt6 (it was already deprecated in Qt5).
Time measurements now have to use QElapsedTimer.

QLayout does not have setMargin anymore, but setContentsMargins must be used instead.

The deprecated QDesktopWidget is no longer available in Qt6 but QScreen has to be used instead.